move NewDeveloperField components to Developers namespace
