--- test/components/new_developer_fields_component_test.rb
@@ -1,36 +1,0 @@
-require "test_helper"
-
-class NewDeveloperFieldsComponentTest < ViewComponent::TestCase
-  test "renders if the user's developer profile is missing fields" do
-    developers(:one).update!(available_on: nil)
-    render_inline NewDeveloperFieldsComponent.new(users(:developer))
-    assert_text I18n.t("new_developer_fields_component.title")
-  end
-
-  test "doesn't render if the developer profile is filled in" do
-    render_inline NewDeveloperFieldsComponent.new(users(:developer))
-    refute_component_rendered
-  end
-
-  test "doesn't render if the user doesn't have a developer profile" do
-    render_inline NewDeveloperFieldsComponent.new(users(:empty))
-    refute_component_rendered
-  end
-
-  test "doesn't render if the developer hasn't been saved yet" do
-    user = users(:empty)
-    user.build_developer
-    render_inline NewDeveloperFieldsComponent.new(user)
-    refute_component_rendered
-  end
-
-  test "doesn't render if the user isn't signed in" do
-    render_inline NewDeveloperFieldsComponent.new(nil)
-    refute_component_rendered
-  end
-
-  test "doesn't render if disabled" do
-    render_inline NewDeveloperFieldsComponent.new(users(:developer), enabled: false)
-    refute_component_rendered
-  end
-end
